% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benr2.R
\docType{package}
\name{benr2-package}
\alias{benr2-package}
\title{The most incredible editable benr2 Package}
\description{
Random functions for dice and math functions.
}
\details{
rolling dice, shuffling decks
}
\section{My Custom Section}{
 Add additional information here.
}
\keyword{package}

