% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4010-Project1.R
\name{roll}
\alias{roll}
\title{roll}
\usage{
roll(dietype = 1:6, probvalues = rep(1/length(dietype), length(dietype)))
}
\arguments{
\item{dietype}{values for die}

\item{probvalues}{value of each possible occurence}
}
\value{
sum of a roll of the dice
}
\description{
roll
}
\examples{
roll()
roll(1:12)
roll(1:6, probvalues = c(rep(1/7,5), 2/7))
}

