% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4010-Project1.R
\name{RP}
\alias{RP}
\title{RP}
\usage{
RP(vector, power)
}
\arguments{
\item{vector}{base number of exponent operation}

\item{power}{power that base is raised  to}
}
\value{
result of the vector being raised to power
}
\description{
RP
}
\examples{
RP(1,0)
RP(3,5)
RP(5,3)
}

