% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4010-Project1.R
\name{shuffle}
\alias{shuffle}
\title{shuffle}
\usage{
shuffle(deck)
}
\arguments{
\item{deck}{vector of numbers to be randomly shuffled}
}
\value{
shuffled vector of numbers (numbers in random order)
}
\description{
shuffle
}
\examples{
shuffle(DECK)
}

