% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSetDocumentation.R
\docType{data}
\name{DECK}
\alias{DECK}
\title{Deck of Cards}
\format{A data frame with 3 variables and 52 rows.}
\usage{
DECK
}
\description{
Data representing a deck of cards.
}
\keyword{dataset}

