% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4010-Project1.R
\name{shuffle}
\alias{shuffle}
\title{shuffle}
\usage{
shuffle(deck)
}
\arguments{
\item{a}{vector of numbers to be randomly shuffled}
}
\value{
shuffled vector of numbers (numbers in random order)
}
\description{
shuffle
}
\examples{
deck <- read.csv("/var/folders/h9/44pdnt9j3n727vp2q8pswr_00000gp/T//RtmpKaNS4I/data22675300904f", stringsAsFactors=FALSE)
deck
shuffle(deck)
}

